+++ codegen/example/src-gen/main/java/org/fuin/ddd4j/codegen/example/TheRootCreateCommand.java
@@ -4,7 +4,6 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.json.bind.annotation.JsonbProperty;
 import jakarta.json.bind.annotation.JsonbTypeAdapter;
 import javax.annotation.Nullable;
-import org.fuin.ddd4j.core.DomainEventExpectedEntityIdPath;
 import org.fuin.ddd4j.core.EntityIdPath;
 import org.fuin.ddd4j.core.EventType;
 import org.fuin.cqrs4j.jsonb.AbstractAggregateCommand;
@@ -37,7 +36,6 @@ import org.fuin.objects4j.ui.Examples;
  */
 @Immutable
 @HasSerializedDataTypeConstant
-@DomainEventExpectedEntityIdPath({ TheRootId.class, TheEntityId.class })
 public final class TheRootCreateCommand extends AbstractAggregateCommand<TheRootId, TheEntityId> {
 
     @Serial