+++ codegen/processor/src/main/java/org/fuin/ddd4j/codegen/processor/CommandVOTemplate.java
@@ -35,6 +35,11 @@ public final class CommandVOTemplate extends AbstractEventVOTemplate<CommandVO> 
     }
 
     @Override
+    public String targetPackage(CommandVO anno) {
+        return anno.pkg();
+    }
+
+    @Override
     protected String getTemplateName() {
         return "CommandVO.java";
     }